backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
xgboost==2.0.3
lightgbm==4.5.0
shap==0.45.1
torch==2.2.2
reportlab==4.2.2
python-dotenv==1.0.1
